Move Error type from zendesk.go to error.go
